# Refactor: Simplify to MVP while enhancing functionality