--- a/requirements.txt
+++ b/requirements.txt
@@ -1,29 +1,8 @@
-# Core packages
-xclim>=0.48.0
-xarray>=2024.1.0
-dask[complete]>=2024.1.0
-netCDF4>=1.6.0
-h5netcdf>=1.2.0
-
-# Raster data handling
-rasterio>=1.3.0
-rioxarray>=0.15.0
-pyproj>=3.6.0
-
-# Data processing
-numpy>=1.24.0
+# Core dependencies for xclim-timber MVP
+xarray>=2023.0.0
 pandas>=2.0.0
-scipy>=1.10.0
+numpy>=1.24.0
+netCDF4>=1.6.0
 
-# Utilities
-pyyaml>=6.0
-tqdm>=4.65.0
-click>=8.1.0
-
-# Optional: for visualization
-matplotlib>=3.7.0
-cartopy>=0.22.0
-
-# Optional: for performance monitoring
-dask-jobqueue>=0.8.0
-distributed>=2024.1.0
+# Optional but recommended for better performance
+bottleneck>=1.3.0  # Speeds up xarray operations