… PR_DESCRIPTION.md
… README.md
… config_simple.yaml
… efficient_extraction.py
… fast_point_extraction.py
… point_extraction.py
… run_test.sh
… src/config.py
… src/config_noresm.py
… src/data_loader.py
… src/indices_calculator.py
… src/pipeline.py
… src/preprocessor.py
… test_annual_temp.py
… visualize_temp.py